Sorts skills and jobs based on search & resume
This file prioritizes the jobs that fit the search the best, jobs that fit both the search and the user's resume the best, skills that fit the jobs that popped up after the search, and skills that fit the search and the user's resume. For further description, please refer to the comments in this coding file. Currently, this file completely runs correctly, but needs to be connected to the database and the file upload.